Improve Resource file tests
