--- tests/Immediate.Validations.FunctionalTests/Common/IsolatedTestOrderer.cs
@@ -1,10 +1,0 @@
-﻿using Xunit;
-using Xunit.Abstractions;
-
-namespace Immediate.Validations.FunctionalTests.Common;
-
-public sealed class IsolatedTestOrderer : ITestCollectionOrderer
-{
-	public IEnumerable<ITestCollection> OrderTestCollections(IEnumerable<ITestCollection> testCollections) =>
-		testCollections.OrderBy(tc => tc.DisplayName == nameof(IsolatedTestCollectionDefinition));
-}
