--- tests/Immediate.Validations.FunctionalTests/xunit.runner.json
@@ -1,4 +1,0 @@
-﻿{
-	"collectionBehavior": "CollectionPerClass",
-	"collectionOrderer": "Immediate.Validations.FunctionalTests.Common.IsolatedTestOrderer, Immediate.Validations.FunctionalTests"
-}
